Fix sign-compare warnings in examples.

+++ drake/examples/kuka_iiwa_arm/run_kuka_iiwa_arm_dynamics.cc
@@ -105,7 +105,8 @@ int main(int argc, char* argv[]) {
   int num_velocities = rigid_body_sys->number_of_velocities();
 
   // Ensures the size of the output is correct.
-  if (final_robot_state.size() != rigid_body_sys->getNumOutputs()) {
+  if (final_robot_state.size() !=
+      static_cast<int>(rigid_body_sys->getNumOutputs())) {
     throw std::runtime_error(
         "ERROR: Size of final robot state (" +
         std::to_string(final_robot_state.size()) +
@@ -133,7 +134,8 @@ int main(int argc, char* argv[]) {
 
   // Ensures the robot's joints are within their position limits.
   std::vector<std::unique_ptr<RigidBody>>& bodies = tree->bodies;
-  for (int robot_state_index = 0, body_index = 0; body_index < bodies.size();
+  for (int robot_state_index = 0, body_index = 0;
+       body_index < static_cast<int>(bodies.size());
        ++body_index) {
     // Skips rigid bodies without a parent (this includes the world link).
     if (!bodies[body_index]->hasParent()) continue;
